+++ ui/app/templates/components/secret-edit.hbs
@@ -34,41 +34,13 @@
   {{/unless}}
   <ToolbarActions>
     {{#if (eq mode 'show')}}
-      {{#if (and canUndeleteVersion (or this.modelForData.deleted this.modelForData.destroyed))}}
-        <button
-            type="button"
-            class="toolbar-link"
-            onclick={{action "handleDelete" "undelete"}}
-          >
-          Undelete
-        </button>
-      {{/if}}
-      {{#if (and (not this.modelForData.deleted) (not this.modelForData.destroyed)) }}
-        {{#if canDestroyVersion}}
-          <button
-              type="button"
-              class="toolbar-link"
-              onclick={{action (mut this.showDeleteModal) true}}
-            >
-            {{if (and (not this.modelForData.deleted) (not this.modelForData.destroyed)) "Delete" "Destroy"}}
-          </button>
-          <div class="toolbar-separator"/>
-        {{else}}
-            {{#if (or canDeleteAnyVersion (and isLatestVersion canDelete))}}
-              <ConfirmAction
-                @buttonClasses="toolbar-link"
-                @confirmTitle="Delete"
-                @confirmMessage="This version will no longer be able to be read, but the underlying data will not be removed and can still be undeleted."
-                @onConfirmAction={{action "handleDelete" (if (and canDelete (not canUpdate)) "delete-latest-version" "delete") }}
-                data-test-secret-v2-delete="true"
-              >
-                Delete
-              </ConfirmAction>
-          {{/if}}
-        {{/if}}
-      {{/if}}  
+      <SecretDeleteMenu 
+        @modelForData={{this.modelForData}}
+        @model={{this.model}}
+        @canDelete={{canDelete}}
+        @canUpdate={{canUpdate}}
+      />
     {{/if}}
-
     {{#if (and (eq mode 'show') (or canEditV2Secret canUpdate))}}
       {{#let (concat 'vault.cluster.secrets.backend.' (if (eq mode 'show') 'edit' 'show')) as |targetRoute|}}
         {{#unless (and isV2 (or isWriteWithoutRead modelForData.destroyed modelForData.deleted))}}
@@ -162,83 +134,5 @@
     {{/if}}
   </ToolbarActions>
 </Toolbar>
-
-<Modal
-  @title="Delete Secret?"
-  @onClose={{action (mut this.showDeleteModal) false}}
-  @isActive={{this.showDeleteModal}}
-  @type="warning"
-  @showCloseButton={{true}}
->
-  <section class="modal-card-body">
-    <p class="has-bottom-margin-s">There are three ways to delete or destroy the <strong>{{this.model.id}}</strong> secret. Each is different, so be sure to read the below carefully.</p>
-    <p class="has-bottom-margin-s"><strong>How would you like to proceed?</strong></p>
-    {{#unless this.modelForData.destroyed}}
-      {{#unless this.modelForData.deleted}}
-        <div class="modal-radio-button">
-          <RadioButton
-            @value="delete"
-            @radioClass="radio"
-            @groupValue={{this.deleteType}}
-            @changed={{action (mut this.deleteType)}}
-            @name="setup-deleteType"
-            @radioId="delete-version"
-          />
-          <div class="helper-text">
-            <label for="dr" data-test-replication-type-select="dr"><strong>Delete this version</strong></label>
-            <p>This deletes Version {{this.modelForData.version}} of the secret. It can be un-deleted later.</p>
-          </div>
-        </div>
-      {{/unless}}
-      <div class="modal-radio-button">
-        <RadioButton
-          @value="destroy"
-          @radioClass="radio"
-          @groupValue={{this.deleteType}}
-          @changed={{action (mut this.deleteType)}}
-          @name="setup-deleteType"
-          @radioId="destroy-version"
-        />
-        <div class="helper-text">
-          <label for="dr" data-test-replication-type-select="dr"><strong>Destroy this version</strong></label>
-          <p>Version {{this.modelForData.version}} is permanently destroyed and cannot be read or recovered later.</p>
-        </div>
-      </div>
-    {{/unless}}
-    {{#if canDestroyAllVersions}}
-      <div class="modal-radio-button">
-        <RadioButton
-          @value="destroy-all-versions"
-          @radioClass="radio"
-          @groupValue={{this.deleteType}}
-          @changed={{action (mut this.deleteType)}}
-          @name="setup-deleteType"
-          @radioId="destroy-all-versions"
-        />
-        <div class="helper-text">
-          <label for="dr" data-test-replication-type-select="dr"><strong>Destroy all versions</strong></label>
-          <p>All secret versions and metadata are permanently destroyed and cannot be read or recovered later.</p>
-        </div>
-      </div>
-    {{/if}}
-  </section>
-  <footer class="modal-card-foot modal-card-foot-outlined">
-    <button
-      type="button"
-      class="button has-text-danger"
-      onclick={{action "handleDelete" this.deleteType}}
-      disabled={{if this.deleteType false true}}
-    >
-      Delete
-    </button>
-    <button
-      type="button"
-      class="button is-secondary"
-      onclick={{action (mut this.showDeleteModal) false}}
-    >
-      Cancel
-    </button>
-  </footer>
-</Modal>
 
 {{partial partialName}}
