+++ ui/app/templates/components/secret-version-menu.hbs
@@ -22,7 +22,6 @@
             {{#if (has-block)}}
               {{yield}}
             {{/if}}
-        {{!-- ARG TODO Add permissions here --}}
             {{#each (reverse @model.versions) as |secretVersion|}}
               <li class="action">
                 <LinkTo class="link" @params={{array (query-params version=secretVersion.version)}} @invokeAction={{action D.actions.close}} >
